chore: eliminate all compiler and clippy warnings (554 → 0)
- cargo check: 0 warnings (was ~100)
- cargo clippy --all-targets: 0 warnings (was 554)
- Mechanical fixes: vec! → slice, clone closure → cloned(), map_or simplification
- Dead code: #[allow(dead_code)] for infrastructure modules, reconcile_all fallbacks
- Unused imports/variables removed across all crates
- assert_eq!(x, true) → assert!(x) in test files
- strip_prefix/strip_suffix instead of manual slicing
- Test modules moved to end of files (items after test module)
- WorkQueue: added Default impl and is_empty()
- Large enum variants: #[allow(clippy::large_enum_variant)] for K8s API types
- Too many arguments: #[allow(clippy::too_many_arguments)] where needed
- Doc comment formatting fixes

+++ crates/api-server/src/handlers/proxy.rs
@@ -307,7 +307,7 @@ pub async fn proxy_service(
                     } else {
                         ports.first()
                     } {
-                        resolved_port = ep_port.port as u16;
+                        resolved_port = ep_port.port;
                     }
                 }
